extended Owner from base, split controllers using service
