+++ backend/src/main/java/com/ted/DeepInThought/controller/OwnerController.java
@@ -1,33 +1,42 @@
 package com.ted.DeepInThought.controller;
 
 import com.ted.DeepInThought.model.Owner;
-import com.ted.DeepInThought.repository.OwnerRepository;
+import com.ted.DeepInThought.service.OwnerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 public class OwnerController {
 
     @Autowired
-    private OwnerRepository ownerRepo;
+    private OwnerService ownerService;
 
     @GetMapping("/owner")
     public ResponseEntity<List<Owner>> getAllOwners() {
         try {
-            List<Owner> ownerList = new ArrayList<>();
-            ownerRepo.findAll().forEach(ownerList::add);
+            List<Owner> ownerList = ownerService.getAllOwners();
 
             if (ownerList.isEmpty()) {
                 return new ResponseEntity<>(HttpStatus.NO_CONTENT);
             }
 
             return new ResponseEntity<>(ownerList, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PostMapping("/owner")
+    public ResponseEntity<Owner> addOwner(@RequestBody Owner newOwner) {
+        try {
+            return new ResponseEntity<>(ownerService.addOwner(newOwner), HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
